refactor: pass state up to crossword component
