+++ client/components/crosswordComponent.tsx
@@ -1,67 +1,27 @@
-import React, { useEffect, useState } from 'react'
-
-import { SquareComponent, SquareComponentProps } from './squareComponent'
-import Crossword from '../types/crossword'
-import { Orientation, Square, SquarePosition, WordPosition } from '../types/types'
+import React, { useState } from "react"
+import Crossword from "../types/crossword"
+import { Square } from "../types/types"
+import { BoardComponent } from "./boardComponent"
 
 interface CrosswordComponentProps {
     crossword: Crossword
 }
 
-export const CrosswordComponent = ({crossword}: CrosswordComponentProps): JSX.Element => {
-    const [selectedSquare, setSelectedSquare] = useState<null | Square>(null) // this state might be passed up
-    const [squaresInSelectedWord, setSquaresInSelectedWord] = useState<Square[]>([])
+export const CrosswordComponent = ({ crossword }: CrosswordComponentProps): JSX.Element => {
+    const [selectedSquare, setSelectedSquare] = useState<null | Square>(null)
 
-    useEffect(() => {
-        if (selectedSquare != null) {
-            const selectedHorizontalWordPosition: WordPosition | undefined = crossword.getHorizontalWord(selectedSquare)?.position
-            const selectedVerticalWordPosition: WordPosition | undefined = crossword.getVerticalWord(selectedSquare)?.position
-            
-            let newSquaresInSelectedWord: Square[] = []
-            if (selectedHorizontalWordPosition != null) {
-                newSquaresInSelectedWord = [...newSquaresInSelectedWord, ...newSquaresInSelectedWord.concat(crossword.wordPositionToSquares(selectedHorizontalWordPosition))]
-            }
-            if (selectedVerticalWordPosition != null) {
-                newSquaresInSelectedWord = [...newSquaresInSelectedWord, ...newSquaresInSelectedWord.concat(crossword.wordPositionToSquares(selectedVerticalWordPosition))]
-            }
-            setSquaresInSelectedWord(newSquaresInSelectedWord)
-        }
-    }, [selectedSquare])
-
-    const handleClickGenerator = (squarePosition: Square): React.MouseEventHandler<HTMLDivElement> => {
+    const handleClickSquare = (square: Square): React.MouseEventHandler<HTMLDivElement> => {
         return (event) => {
             event.preventDefault()
-            setSelectedSquare(squarePosition)
+            setSelectedSquare(square)
         }
     }
 
-    const SQUARE_WIDTH = 75
-    const SQUARE_HEIGHT = 75
-
-    const squareComponents: Array<React.ReactElement<SquareComponentProps, any>> = crossword.squareArray.flat().map(
-        (square) => {
-            const isInSelectedWord: boolean = squaresInSelectedWord == null ? false : squaresInSelectedWord.reduce<boolean>((prev, cur) => {
-                return prev || (cur.position.x === square.position.x && cur.position.y === square.position.y)
-            }, false)
-            return (
-                <SquareComponent
-                    key={(square.position.x << 13) + square.position.y}
-                    square={square}
-                    width={SQUARE_WIDTH}
-                    height={SQUARE_HEIGHT}
-                    handleClick={handleClickGenerator(square)}
-                    isSelected={selectedSquare != null && selectedSquare.position.x === square.position.x && selectedSquare.position.y === square.position.y}
-                    isInSelectedWord={isInSelectedWord}
-                />
-            )
-        }
-    )
-
     return (
-        <>
-            <div className={`w-[${crossword.width * SQUARE_WIDTH}px] h-[${crossword.height * SQUARE_HEIGHT}px] grid grid-cols-${crossword.width} grid-rows-${crossword.height}`}>
-                {squareComponents}
-            </div>
-        </>
+        <BoardComponent
+            crossword={crossword}
+            selectedSquare={selectedSquare}
+            handleClickSquare={handleClickSquare}
+        />
     )
 }
